#76 Rename module (parser → datamodel)
