+++ xarray_dataclasses/dataarray.py
@@ -14,7 +14,7 @@ from typing_extensions import Literal, ParamSpec, Protocol
 
 
 # submodules
-from .parser import DataModel
+from .datamodel import DataModel
 from .utils import copy_class
